refactor: rename directories
