+++ jobs/Frontend/fe-challenge/src/pages/Search/Search.test.tsx
@@ -8,7 +8,7 @@ import {
   waitFor,
   within,
 } from '@/utils/test-utils';
-import Search from '@/pages/Search';
+import Search from '@/pages/search';
 import { server } from '@/mocks/server';
 import { routes } from '@/router';
 import { moviesResultMock } from '@/mocks/data';
