+++ jobs/Frontend/fe-challenge/src/pages/Search/components/SearchResult.tsx
@@ -1,10 +1,10 @@
 import React, { useMemo } from 'react';
-import MovieCard from '@/pages/Search/components/MovieCard';
-import useSearchMovie from '@/pages/Search/hooks/useSearchMovie';
-import EmptySearch from '@/pages/Search/components/EmptySearch';
-import useInfiniteScroll from '@/pages/Search/hooks/useInfiniteScroll';
-import ResultsNumber from '@/pages/Search/components/ResultsNumber';
-import usePrefetchMovieDetail from '@/pages/MovieDetail/hooks/usePrefetchMovieDetail';
+import MovieCard from '@/pages/search/components/MovieCard';
+import useSearchMovie from '@/pages/search/hooks/useSearchMovie';
+import EmptySearch from '@/pages/search/components/EmptySearch';
+import useInfiniteScroll from '@/pages/search/hooks/useInfiniteScroll';
+import ResultsNumber from '@/pages/search/components/ResultsNumber';
+import usePrefetchMovieDetail from '@/pages/movie-detail/hooks/usePrefetchMovieDetail';
 
 interface SearchResultProps {
   query: string;
