+++ jobs/Frontend/fe-challenge/src/pages/error-page/index.ts
