+++ jobs/Frontend/fe-challenge/src/pages/movie-detail/components/MovieActors.tsx
