+++ jobs/Frontend/fe-challenge/src/pages/movie-detail/components/MovieGenres.tsx
