+++ jobs/Frontend/fe-challenge/src/router.tsx
@@ -1,8 +1,8 @@
 import { Outlet, createBrowserRouter } from 'react-router-dom';
 import Layout from '@/components/Layout';
-import ErrorPage from '@/pages/ErrorPage';
-import Search from '@/pages/Search';
-import MovieDetail from '@/pages/MovieDetail';
+import ErrorPage from '@/pages/error-page';
+import Search from '@/pages/search';
+import MovieDetail from '@/pages/movie-detail';
 
 export const routes = [
   {
